cleaning up other variations
