+++ template/src/components/Controls.tsx
@@ -57,9 +57,6 @@ import {useScreenshare} from '../subComponents/screenshare/useScreenshare';
 import LayoutIconDropdown from '../subComponents/LayoutIconDropdown';
 import TranscriptIcon from '../../src/subComponents/caption/TranscriptIcon';
 import CaptionIcon from '../../src/subComponents/caption/CaptionIcon';
-import CaptionIcon2 from '../../src/subComponents/caption/CaptionIcon2';
-import CaptionIcon3 from '../../src/subComponents/caption/CaptionIcon3';
-import CaptionIcon4 from '../../src/subComponents/caption/CaptionIcon4';
 
 const MoreButton = () => {
   const {rtcProps} = useContext(PropsContext);
@@ -313,21 +310,9 @@ const Controls = () => {
             <CopyJoinInfo />
           </View>
 
+          {/* TODO:// to be refactored */}
           <View testID="caption-btn" style={{marginHorizontal: 10}}>
             <CaptionIcon />
-          </View>
-          {/* TODO:to be removed */}
-          <View testID="caption-btn" style={{marginHorizontal: 10}}>
-            <CaptionIcon2 />
-          </View>
-
-          {/* TODO:to be removed */}
-          <View testID="caption-btn" style={{marginHorizontal: 10}}>
-            <CaptionIcon3 />
-          </View>
-          {/* TODO:to be removed */}
-          <View testID="caption-btn" style={{marginHorizontal: 10}}>
-            <CaptionIcon4 />
           </View>
 
           <View testID="transcript-btn" style={{marginHorizontal: 10}}>
