+++ template/src/subComponents/caption/useCaption.tsx
@@ -8,54 +8,40 @@ interface Transcript {
 export const CaptionContext = React.createContext<{
   isCaptionON: boolean;
   setIsCaptionON: React.Dispatch<React.SetStateAction<boolean>>;
-  isCaptionON2: boolean;
-  setIsCaptionON2: React.Dispatch<React.SetStateAction<boolean>>;
-  isCaptionON3: boolean;
-  setIsCaptionON3: React.Dispatch<React.SetStateAction<boolean>>;
-  isCaptionON4: boolean;
-  setIsCaptionON4: React.Dispatch<React.SetStateAction<boolean>>;
   transcript: Transcript;
   setTranscript: React.Dispatch<React.SetStateAction<Transcript>>;
   isSTTActive: boolean;
   setIsSTTActive: React.Dispatch<React.SetStateAction<boolean>>;
+  meetingTranscript: [];
+  setMeetingTranscript: React.Dispatch<React.SetStateAction<[]>>;
 }>({
   isCaptionON: false,
   setIsCaptionON: () => {},
-  isCaptionON2: false,
-  setIsCaptionON2: () => {},
-  isCaptionON3: false,
-  setIsCaptionON3: () => {},
-  isCaptionON4: false,
-  setIsCaptionON4: () => {},
   transcript: {},
   setTranscript: () => {},
   isSTTActive: false,
   setIsSTTActive: () => {},
+  meetingTranscript: [],
+  setMeetingTranscript: () => {},
 });
 
 const CaptionProvider = ({children}) => {
   const [isCaptionON, setIsCaptionON] = React.useState<boolean>(false);
-  const [isCaptionON2, setIsCaptionON2] = React.useState<boolean>(false); //TODO: to be removed
-  const [isCaptionON3, setIsCaptionON3] = React.useState<boolean>(false); //TODO: to be removed
-  const [isCaptionON4, setIsCaptionON4] = React.useState<boolean>(false); //TODO: to be removed
   const [transcript, setTranscript] = React.useState<Transcript>({});
   const [isSTTActive, setIsSTTActive] = React.useState<boolean>(false);
+  const [meetingTranscript, setMeetingTranscript] = React.useState([]);
 
   return (
     <CaptionContext.Provider
       value={{
         isCaptionON,
         setIsCaptionON,
-        isCaptionON2,
-        setIsCaptionON2,
-        isCaptionON3,
-        setIsCaptionON3,
-        isCaptionON4,
-        setIsCaptionON4,
         transcript,
         setTranscript,
         isSTTActive,
         setIsSTTActive,
+        meetingTranscript,
+        setMeetingTranscript,
       }}>
       {children}
     </CaptionContext.Provider>
